Made all the methods synchronized
This seems like a stupid and brutal approach, but really it kinda
makes sense. All the operations there should be atomic so doing
this guarantees that. There might be smaller increments of
synchronization that might be possible but I do not think that
it might be worth finding them. This does not appear to be an
intensive application, so simply guaranteeing correctness should
suffice.